Update package dependencies to version 20250902.1546.20 across multiple modules
This commit updates the versions of several @tmlmobilidade packages in the package.json and package-lock.json files for the alerts and auth applications. The changes ensure consistency in dependency versions, enhancing compatibility and stability across the project. Additionally, it updates the version of nodemailer and nuqs to their latest versions, improving functionality and security.