Add GitHub Action for module discovery
This commit introduces a new GitHub Action that discovers which modules and apps to build based on changed files. The action takes a newline-separated list of modules and their associated apps as input and outputs a JSON array of modules to build along with a mapping of module names to their apps. Additionally, the existing workflow for building all modules has been updated to utilize this new action, streamlining the process of detecting changes and determining the necessary builds.